+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added an option to allow all users/guests to view RSS
 - Added studio-viewPublishedRSS-[PodcastUID] permission to check accessing to RSS
 - Added checkAccessToVolumes general setting which checks saveAssets:[VolumeUID] permission before listing volumes on asset index page
+- Added lastBuildDate for podcast RSS
 - Now each podcast has two settings pages (general settings, episode settings)
 - Improved fetch from RSS via CURL 
 - Improved fetch from RSS progress notifications